@@ -2,15 +2,5 @@
 
 module.exports = {
   env: { node: true },
-  extends: 'standard',
-  globals: {
-    And: true,
-    After: true,
-    Before: true,
-    Feature: true,
-    Given: true,
-    Scenario: true,
-    Then: true,
-    When: true
-  }
+  extends: 'standard'
 }
